Added DNSChecker
Does a dns request on each domain, and removes the domains that aren't registered from the blocklist. It reduces blocklist size, which hopefully would increase the speed of blockers.

The results are cached for 7 days to reduce server load and creation of list time, and only 7 days to ensure if domains are later reregistered then they will be blocked again